// deploymentcli/web/src/utils/local-client/types/openapi.yml/providerV2.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * Example API
 * Example API
 * OpenAPI spec version: 1.0
 */
import type { ProviderV2Status } from './providerV2Status';

/**
 * ProviderV2
 */
export interface ProviderV2 {
  name: string;
  team: string;
  version: string;
  stackId: string;
  status: ProviderV2Status;
  type: string;
  id: string;
  alias: string;
  functionArn?: string;
  functionRoleArn?: string;
}